+++ assets/javascripts/kitten/components/navigation/header-nav/components/user-menu.js
@@ -1,6 +1,6 @@
 import React, { useRef } from 'react'
 import PropTypes from 'prop-types'
-import Dropdown from '../../../../components/dropdowns/v2/dropdown'
+import { Dropdown } from './dropdown'
 import { Context } from './context'
 import { getReactElementsByType } from '../../../../helpers/react/react-elements'
 import classNames from 'classnames'